# Conflicts:
#	modules/core/src/test/java/org/apache/ignite/testsuites/IgniteCacheTestSuite5.java

+++ modules/ml/src/test/java/org/apache/ignite/ml/dataset/primitive/DatasetWrapperTest.java
@@ -61,7 +61,7 @@ public class DatasetWrapperTest {
     public void testComputeWithCtx() {
         doReturn(42).when(dataset).computeWithCtx(any(IgniteTriFunction.class), any(), any());
 
-        Integer res = wrapper.computeWithCtx(mock(IgniteTriFunction.class), mock(IgniteBinaryOperator.class),
+        Integer res = (Integer) wrapper.computeWithCtx(mock(IgniteTriFunction.class), mock(IgniteBinaryOperator.class),
             null);
 
         assertEquals(42, res.intValue());
@@ -75,7 +75,7 @@ public class DatasetWrapperTest {
     public void testComputeWithCtx2() {
         doReturn(42).when(dataset).computeWithCtx(any(IgniteTriFunction.class), any(), any());
 
-        Integer res = wrapper.computeWithCtx(mock(IgniteBiFunction.class), mock(IgniteBinaryOperator.class),
+        Integer res = (Integer) wrapper.computeWithCtx(mock(IgniteBiFunction.class), mock(IgniteBinaryOperator.class),
             null);
 
         assertEquals(42, res.intValue());
@@ -102,7 +102,8 @@ public class DatasetWrapperTest {
     public void testCompute() {
         doReturn(42).when(dataset).compute(any(IgniteBiFunction.class), any(), any());
 
-        Integer res = wrapper.compute(mock(IgniteBiFunction.class), mock(IgniteBinaryOperator.class), null);
+        Integer res = (Integer) wrapper.compute(mock(IgniteBiFunction.class), mock(IgniteBinaryOperator.class),
+            null);
 
         assertEquals(42, res.intValue());
 
@@ -115,7 +116,8 @@ public class DatasetWrapperTest {
     public void testCompute2() {
         doReturn(42).when(dataset).compute(any(IgniteBiFunction.class), any(IgniteBinaryOperator.class), any());
 
-        Integer res = wrapper.compute(mock(IgniteFunction.class), mock(IgniteBinaryOperator.class), null);
+        Integer res = (Integer) wrapper.compute(mock(IgniteFunction.class), mock(IgniteBinaryOperator.class),
+            null);
 
         assertEquals(42, res.intValue());
